requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cycler==0.10.0
Django==3.0.7
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-silk==4.1.0
djangorestframework==3.11.0
drf-extra-fields==3.0.3
drf-yasg==1.17.1
gprof2dot==2019.11.30
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mysql-connector-python==8.0.20
mysqlclient==1.4.6
numpy==1.19.2
openapi-codec==1.3.2
packaging==20.4
Pillow==8.0.1
protobuf==3.12.2
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
